requirements.txt
apache-airflow==2.7.1
apache-airflow-providers-amazon==9.6.1
apache-airflow-providers-postgres==6.1.3
requests==2.32.3
boto3>=1.26.0
psycopg2-binary>=2.9.5
